Fix pssg model import for shaders with >65535 vertices

+++ src/EgoEngineLibrary/Formats/Pssg/GltfCarExteriorPssgConverter.cs
@@ -355,9 +355,9 @@ namespace EgoEngineLibrary.Formats.Pssg
                     var b = tri.B + baseVertexIndex;
                     var c = tri.C + baseVertexIndex;
 
-                    rds.Indices.Add((ushort)a);
-                    rds.Indices.Add((ushort)b);
-                    rds.Indices.Add((ushort)c);
+                    rds.Indices.Add((uint)a);
+                    rds.Indices.Add((uint)b);
+                    rds.Indices.Add((uint)c);
                 }
 
                 // Add the matrixpalletejointnode id to the shader's joint list
